Reorganize collections folder (#440)
* reorganize collections folder

* update podspec
