--- SkeletonViewCore/Sources/Collections/Generics/GenericCollectionView.swift
@@ -1,13 +1,0 @@
-//
-//  GenericCollectionView.swift
-//  SkeletonView-iOS
-//
-//  Created by Juanpe Catalán on 30/03/2018.
-//  Copyright © 2018 SkeletonView. All rights reserved.
-//
-
-import UIKit
-
-protocol GenericCollectionView {
-    var scrollView: UIScrollView { get }
-}
