+++ SkeletonViewCore/Sources/Internal/Collections/CollectionSkeleton.swift
@@ -14,6 +14,7 @@ enum CollectionAssociatedKeys {
 }
 
 protocol CollectionSkeleton {
+    
     var skeletonDataSource: SkeletonCollectionDataSource? { get set }
     var skeletonDelegate: SkeletonCollectionDelegate? { get set }
     var estimatedNumberOfRows: Int { get }
@@ -23,9 +24,11 @@ protocol CollectionSkeleton {
     func removeDummyDataSource(reloadAfter: Bool)
     func disableUserInteraction()
     func enableUserInteraction()
+    
 }
 
 extension CollectionSkeleton where Self: UIScrollView {
+    
     var estimatedNumberOfRows: Int { return 0 }
     func addDummyDataSource() {}
     func removeDummyDataSource(reloadAfter: Bool) {}
@@ -43,4 +46,5 @@ extension CollectionSkeleton where Self: UIScrollView {
             isScrollEnabled = true
         }
     }
+    
 }
