+++ SkeletonViewCore/Sources/Internal/Collections/SkeletonCollectionDelegate.swift
